chek creat card

+++ index.js
@@ -78,7 +78,7 @@ const getBalance = async (valueId) => {
   return totalAmount;
 };
 const setBalance = async (Amount, Description, userId, cardDate) => {
-  var parsedAmount = parseFloat(Amount, 10);
+  var parsedAmount = parseInt(Amount, 10);
   //const monthlyExpenseFake = 'a012w000002UCygAAG';
   await client.query(`INSERT INTO salesforce.Expense_Card__c(sfid, Amount__c, Card_Keeper__c, Card_Date__c, Description__c, ExterId__c)
   VALUES('${userId}', ${parsedAmount}, '${userId}', '${cardDate}', '${Description}', gen_random_uuid());`)
